+++ app/mentor/tasks-management/_actions/create-task.ts
@@ -1,36 +1,18 @@
 'use server'
 
+import { revalidatePath } from 'next/cache'
+import { redirect } from 'next/navigation'
 import prisma from '@/lib/prisma'
 import { getCurrentUser } from '@/utils/users'
 import { handleFileSave, handleFileUpload } from '@/utils/fileService'
-import { z } from 'zod'
+import { TaskFormSchema } from '@/utils/task-schema'
 
-const max_Upload_Size = 1024 * 1024 * 5 // 5MB
-const accepted_File_Types = ['application/pdf', 'application/msword']
+const CreateTask = TaskFormSchema.omit({ id: true })
 
-const taskSchema = z.object({
-  title: z.string().min(1, { message: 'Title is required' }),
-  description: z.string().min(1, { message: 'Description is required' }),
-  startDate: z.date(),
-  endDate: z.date(),
-  upload: z
-    .instanceof(File)
-    .optional()
-    .refine(file => {
-      return !file
-    }, 'File is required')
-    .refine(file => {
-      return !file || file.size <= max_Upload_Size
-    }, 'File size must be less than 5MB')
-    .refine(file => {
-      return !file || accepted_File_Types.includes(file.type)
-    }, 'File must be a PDF or Word document'),
-})
-
-export const createTask = async (formData: FormData) => {
+export const createNewTask = async (formData: FormData) => {
   const user = await getCurrentUser()
 
-  const taskData = taskSchema.safeParse({
+  const validateFields = CreateTask.safeParse({
     title: formData.get('title'),
     description: formData.get('description'),
     startDate: new Date(formData.get('startDate') as string),
@@ -38,8 +20,8 @@ export const createTask = async (formData: FormData) => {
     upload: formData.get('upload'),
   })
 
-  if (!taskData.success) {
-    const errorMessages = taskData.error.flatten().fieldErrors
+  if (!validateFields.success) {
+    const errorMessages = validateFields.error.flatten().fieldErrors
     const serverErrors = Object.keys(errorMessages).reduce((errors, key) => {
       const errorMessage = errorMessages[key as keyof typeof errorMessages]
       return errorMessage ? { ...errors, [key]: errorMessage[0] } : errors
@@ -47,29 +29,30 @@ export const createTask = async (formData: FormData) => {
     throw new Error(JSON.stringify(serverErrors))
   }
 
-  const { title, description, startDate, endDate, upload } = taskData.data
+  const { title, description, startDate, endDate, upload } = validateFields.data
 
   const file = upload as File
-  const taskTitle = title
-  const taskDescription = description
-  const taskStartDate = startDate
-  const taskEndDate = endDate
 
   const metadata = { contentType: file.type }
 
-  const newTask = await prisma.task.create({
-    data: {
-      title: taskTitle,
-      description: taskDescription,
-      status: 'PENDING',
-      startDate: taskStartDate,
-      endDate: taskEndDate,
-      mentorId: user?.id || '',
-    },
-  })
-
-  console.log('New task created with ID', newTask.id)
-
-  const downloadURL = await handleFileUpload(file, metadata)
-  await handleFileSave(file, downloadURL, user?.id || '', newTask.id)
+  try {
+    const newTask = await prisma.task.create({
+      data: {
+        title: title,
+        description: description,
+        status: 'PENDING',
+        startDate: startDate,
+        endDate: endDate,
+        mentorId: user?.id || '',
+      },
+    })
+    const downloadURL = await handleFileUpload(file, metadata)
+    await handleFileSave(file, downloadURL, user?.id || '', newTask.id)
+    console.log(`New task created with ID ${newTask.id}`)
+  } catch (error) {
+    throw new Error('Database Error: Failed to Create Task.')
+  } finally {
+    revalidatePath('/mentor/tasks-management')
+    redirect('/mentor/tasks-management')
+  }
 }
